Refactor CLI parts of header.py into annotate.py
